refact: TokenDataStoreManager 인터페이스 생성 및 기존 TokenDataStoreManager class 에 상속
