+++ data/src/main/java/com/prac/data/source/network/di/OkHttpClientModule.kt
@@ -4,7 +4,6 @@ import com.prac.data.BuildConfig
 import com.prac.data.repository.TokenRepository
 import com.prac.data.source.network.di.annotation.AuthOkHttpClient
 import com.prac.data.source.network.di.annotation.BasicOkHttpClient
-import com.prac.data.source.local.datastore.TokenDataStoreManager
 import com.prac.data.source.network.service.AuthorizationInterceptor
 import dagger.Module
 import dagger.Provides
