+++ data/src/main/java/com/prac/data/source/network/service/AuthorizationInterceptor.kt
@@ -1,7 +1,6 @@
 package com.prac.data.source.network.service
 
 import com.prac.data.repository.TokenRepository
-import com.prac.data.source.local.datastore.TokenDataStoreManager
 import kotlinx.coroutines.runBlocking
 import okhttp3.Interceptor
 import okhttp3.Response
